# core/generated/swagger-ui/schema-registry-api-spec.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Confluent Schema Registry"
schemes:
- "http"
- "https"
paths:
...
  /compatibility/subjects/{subject}/versions/{version}:
    post:
      summary: "Test input schema against a particular version of a subject's schema\
        \ for compatibility."
      description: "the compatibility level applied for the check is the configured\
        \ compatibility level for the subject (http:get:: /config/(string: subject)).\
        \ If this subject's compatibility level was never changed, then the global\
        \ compatibility level applies (http:get:: /config)."
      operationId: "testCompatibilityBySubjectName"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "Content-Type"
        in: "header"
        required: false
        type: "string"
      - name: "Accept"
        in: "header"
        required: false
        type: "string"
      - name: "subject"
        in: "path"
        description: "Subject of the schema version against which compatibility is\
          \ to be tested"
        required: true
        type: "string"
        pattern: ".+"
      - name: "version"
        in: "path"
        description: "Version of the subject's schema against which compatibility\
          \ is to be tested. Valid values for versionId are between [1,2^31-1] or\
          \ the string \"latest\".\"latest\" checks compatibility of the input schema\
          \ with the last registered schema under the specified subject"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "Schema"
        required: true
        schema:
          $ref: "#/definitions/RegisterSchemaRequest"
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/CompatibilityCheckResponse"
        404:
          description: "Error code 40401 -- Subject not found\nError code 40402 --\
            \ Version not found"
        422:
          description: "Error code 42201 -- Invalid schema or schema type\nError code\
            \ 42202 -- Invalid version"
        500:
          description: "Error code 50001 -- Error in the backend data store"
...
  /schemas/ids/{id}:
    get:
      summary: "Get the schema string identified by the input ID."
      description: ""
      operationId: "getSchema"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "id"
        in: "path"
        description: "Globally unique identifier of the schema"
        required: true
        type: "integer"
        format: "int32"
      - name: "format"
        in: "query"
        required: false
        type: "string"
      - name: "fetchMaxId"
        in: "query"
        required: false
        type: "boolean"
        default: false
      - name: "Authorization"
        in: "header"
        required: false
        type: "string"
      - name: "Cookie"
        in: "header"
        required: false
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SchemaString"
        404:
          description: "Error code 40403 -- Schema not found\n"
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
...
definitions:
  CompatibilityCheckResponse:
    type: "object"
    properties:
      is_compatible:
        type: "boolean"
...
  RegisterSchemaRequest:
    type: "object"
    properties:
      version:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int32"
      schemaType:
        type: "string"
      references:
        type: "array"
        items:
          $ref: "#/definitions/SchemaReference"
      schema:
        type: "string"
...
  SchemaReference:
    type: "object"
    properties:
      name:
        type: "string"
      subject:
        type: "string"
      version:
        type: "integer"
        format: "int32"
  SchemaString:
    type: "object"
    properties:
      schemaType:
        type: "string"
        description: "Schema type"
      schema:
        type: "string"
        description: "Schema string identified by the ID"
      references:
        type: "array"
        description: "Schema references"
        items:
          $ref: "#/definitions/SchemaReference"
      maxId:
        type: "integer"
        format: "int32"
        description: "Maximum ID"
...